# src/python-backend/api_gateway/main.py
import os
from fastapi import FastAPI, Request
from fastapi.middleware.cors import CORSMiddleware
import httpx
from dotenv import load_dotenv
from fastapi.responses import JSONResponse, Response
import json
from datetime import datetime
import jwt


load_dotenv()

#SERVICE URLS (FROM .env)

ATTENDANCE_SVC = os.getenv("ATTENDANCE_SVC_URL", "http://localhost:8000")
EVENT_SVC = os.getenv("EVENT_SVC_URL", "http://localhost:8020")
LOGIN_SVC = os.getenv("LOGIN_SVC_URL", "http://localhost:8070")
FRONT = os.getenv("FRONT_URL", "http://localhost:8050")


#FASTAPI APP

app = FastAPI(title="API Gateway")

app.add_middleware(
    CORSMiddleware,
    allow_origins=[FRONT, "http://localhost:8050"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["Authorization", "X-Refresh-Token", "Content-Type"],
)

# HELPERS

def is_token_expired(token: str) -> bool:
    """Check if JWT token is expired."""
    try:
        payload = jwt.decode(token, options={"verify_signature": False})
        exp = payload.get("exp")
        return exp * 1000 < datetime.now().timestamp() * 1000
    except:
        return True
    
async def refresh_access_token(refresh_token: str) -> str | None:
    """Call auth service to refresh the access token."""
    try:
        async with httpx.AsyncClient(timeout=30.0) as client:
            resp = await client.post(
                f"{LOGIN_SVC}/auth/refresh",
                json={"refreshToken": refresh_token}
            )
            if resp.status_code == 200:
                data = resp.json()
                return data.get("accessToken")
    except Exception as e:
        print(f"Refresh failed: {e}")
    return None

@app.middleware("http")
async def validate_and_refresh_middleware(request: Request, call_next):
    """
    Function to validate the JWT tokens before each request sent from FrontEnd.
    It also tries to auto-refresh access JWT.
    """

    if request.method == "OPTIONS":
        return await call_next(request)
    
    # Public endpoints that skip auth
    private_paths = ["/attendance/waitlist"]
    public_paths = ["/health", "/auth", "/attendance", "/events/verify/"]
    is_public = any(request.url.path.startswith(prefix) for prefix in public_paths)
    is_private = any(request.url.path.startswith(prefix) for prefix in private_paths)
    if is_public and not is_private:
        return await call_next(request)
    

    refresh_token = request.headers.get("x-refresh-token")
    if not refresh_token or is_token_expired(refresh_token):
        return JSONResponse(
                {"detail": "Invalid refresh token", "code": "LOGOUT_REQUIRED"},
                status_code=401)
    
    access_token = None
    jwt_refreshed = False
    auth_header = request.headers.get("Authorization", "")
    
    if auth_header.startswith("Bearer "):
        access_token = auth_header.split(" ")[1]
    
    if not access_token:
        return JSONResponse({"detail": "Unauthorized"}, status_code=401)
    

    if is_token_expired(access_token): # Second try to refresh Access JWT (Non-expired)
        access_token = await refresh_access_token(refresh_token)
        jwt_refreshed = True
    
    if not access_token:
        return JSONResponse({"detail": "Unauthorized"}, status_code=401)
       
    request.state.access_token = f"Bearer {access_token}"
    
    response = await call_next(request)
    response.headers["Access-Token-Refreshed"] = "true" if jwt_refreshed else "false"
    return response
    

#GENERIC PROXY FUNCTION

async def proxy_request(method: str, url: str, request: Request):
    """Proxy request to backend services with proper error handling."""
    try:
        async with httpx.AsyncClient(timeout=30.0) as client:
            # Get body for POST/PUT requests
            body = None
            if method in ["POST", "PUT"]:
                try:
                    body = await request.json()
                except:
                    pass
            
            # Prepare headers (pass through auth if available)
            headers = {}
            if hasattr(request.state, "access_token"):
                headers["Authorization"] = request.state.access_token
            
            # Make the request
            print(f"🔄 Proxying {method} to {url}")
            print(f"   Body: {body}")
            print(f"   Params: {dict(request.query_params)}")
            
            response = await client.request(
                method,
                url,
                json=body,
                params=dict(request.query_params),
                headers=headers
            )
            
            print(f"   Response status: {response.status_code}")
            
            # Try to parse JSON response
            try:
                return JSONResponse(
                    status_code=response.status_code, 
                    content=response.json()
                )
            except json.JSONDecodeError:
                return JSONResponse(
                    status_code=response.status_code, 
                    content={"message": response.text}
                )
                
    except httpx.ConnectError as e:
        print(f"❌ Connection error to {url}: {e}")
        return JSONResponse(
            status_code=503,
            content={"detail": f"Service unavailable: {url}", "error": str(e)}
        )
    except httpx.TimeoutException as e:
        print(f"⏱ Timeout connecting to {url}: {e}")
        return JSONResponse(
            status_code=504,
            content={"detail": f"Service timeout: {url}", "error": str(e)}
        )
    except Exception as e:
        print(f"❌ Unexpected error proxying to {url}: {e}")
        return JSONResponse(
            status_code=500,
            content={"detail": "Internal server error", "error": str(e)}
        )

#ATTENDANCE ROUTES

@app.post("/attendance/confirm/")
async def attendance_confirm(request: Request):
    return await proxy_request("POST", f"{ATTENDANCE_SVC}/attendance/confirm/", request)

@app.put("/attendance/update/")
async def attendance_update(request: Request):
    return await proxy_request("PUT", f"{ATTENDANCE_SVC}/attendance/update/", request)

@app.get("/attendance/event/{event_id}")
async def attendance_get(event_id: int, request: Request):
    url = f"{ATTENDANCE_SVC}/attendance/event/{event_id}"
    return await proxy_request("GET", url, request)

@app.get("/attendance/check/document/{document_id}/event/{event_id}")
async def attendance_check(document_id: str, event_id: int, request: Request):
    url = f"{ATTENDANCE_SVC}/attendance/check/document/{document_id}/event/{event_id}"
    return await proxy_request("GET", url, request)

@app.put("/attendance/waitlist/switch/id/{document}/event/{event_id}")
async def attendance_waitlist(document: str, event_id: int, request: Request):
    url = f"{ATTENDANCE_SVC}/attendance/waitlist/switch/id/{document}/event/{event_id}"
    return await proxy_request("PUT", url, request)

#EVENT MANAGER ROUTES

@app.post("/events/")
async def create_event(request: Request):
    return await proxy_request("POST", f"{EVENT_SVC}/events/", request)

@app.get("/events/user/{user_id}")
async def events_by_user(user_id: int, request: Request):
    return await proxy_request("GET", f"{EVENT_SVC}/events/user/{user_id}", request)

@app.delete("/events/{event_id}")
async def delete_event(event_id: int, request: Request):
    return await proxy_request("DELETE", f"{EVENT_SVC}/events/{event_id}", request)

@app.get("/events/{event_id}/share")
async def share_event(event_id: int, request: Request):
    return await proxy_request("GET", f"{EVENT_SVC}/events/{event_id}/share", request)

@app.get("/events/verify/{event_id}")
async def verify_event(event_id: int, request: Request):
    return await proxy_request("GET", f"{EVENT_SVC}/events/{event_id}", request)


#LOGIN ROUTES

@app.post("/auth/register")
async def auth_register(request: Request):
    return await proxy_request("POST", f"{LOGIN_SVC}/auth/register", request)

@app.post("/auth/login")
async def auth_login(request: Request):
    auth_resp = await proxy_request("POST", f"{LOGIN_SVC}/auth/login", request)

    if auth_resp.status_code != 200:
        return auth_resp
    
    data = json.loads(auth_resp.body.decode("utf-8"))

    print(data)

    return data

@app.post("/auth/refresh")
async def auth_refresh(request: Request):
    """Refresh access token using httpOnly refresh token."""
    refresh_token = request.headers.get("X-Refresh-Token")
    
    if not refresh_token:
        return JSONResponse({"detail": "No refresh token"}, status_code=401)
    
    try:
        async with httpx.AsyncClient(timeout=30.0) as client:
            resp = await client.post(
                f"{LOGIN_SVC}/auth/refresh",
                json={"refreshToken": refresh_token}
            )
        
        if resp.status_code != 200:
            return JSONResponse({"detail": "Refresh failed"}, status_code=401)
        
        data = resp.json()
        new_access_token = data.get("accessToken")
        
        return JSONResponse(
            {"token": new_access_token, "code": "SUCCESFUL_REFRESH"}, 
            status_code=200
        )
    except Exception as e:
        print(f"❌ Error in auth refresh: {e}")
        return JSONResponse({"detail": "Refresh failed", "error": str(e)}, status_code=500)

@app.post("/auth/userinfo")
async def auth_userinfo(request: Request):
    return await proxy_request("GET", f"{LOGIN_SVC}/auth/userinfo", request)

# HEALTH CHECK

@app.get("/health")
def health():
    return {"gateway": "ok"}


if __name__ == "__main__":
    import uvicorn
    uvicorn.run("main:app", host="0.0.0.0", port=8010)